+++ internal/provider/resource_cert_request.go
@@ -60,18 +60,13 @@ func createCertRequest(_ context.Context, d *schema.ResourceData, _ interface{})
 		return diag.Errorf("error setting value on key 'key_algorithm': %s", err)
 	}
 
-	subjectConfs := d.Get("subject").([]interface{})
-	if len(subjectConfs) != 1 {
-		return diag.Errorf("must have exactly one 'subject' block")
-	}
-	subjectConf, ok := subjectConfs[0].(map[string]interface{})
-	if !ok {
-		return diag.Errorf("subject block cannot be empty")
-	}
-	subject := distinguishedNamesFromSubjectAttributes(subjectConf)
+	// Look for a 'subject' block
+	subject := createSubjectDistinguishedNames(d.Get("subject").([]interface{}))
 
-	certReq := x509.CertificateRequest{
-		Subject: *subject,
+	// Add a `Subject` to the `Certificate` only if it was provided
+	certReq := x509.CertificateRequest{}
+	if subject != nil {
+		certReq.Subject = *subject
 	}
 
 	dnsNamesI := d.Get("dns_names").([]interface{})